Serial-Server/requirements.txt
numpy==2.2.5
pandas==2.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
